# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a management backend API server.",
        "title": "Management BE API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/teams": {
            "get": {
                "description": "Get a paginated list of all teams",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "List all teams",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "pagination": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "items_per_page": {
                                                    "type": "integer"
                                                },
                                                "total_items": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "teams": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/team.TeamResponse"
                                            }
                                        }
                                    }
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new team with name, company name, and contact information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Create a new team",
                "parameters": [
                    {
                        "description": "Team information",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/team.TeamResponse"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teams/{id}": {
            "get": {
                "description": "Get detailed information about a specific team including match statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Get a team by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/team.TeamWithStatsResponse"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing team's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Update a team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Team information",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.UpdateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/team.TeamResponse"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing team by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Delete a team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "team.CreateTeamRequest": {
            "type": "object",
            "required": [
                "company_name",
                "contact_person",
                "name"
            ],
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "FC Barcelona Sports Club"
                },
                "contact_email": {
                    "type": "string",
                    "example": "contact@fcbarcelona.com"
                },
                "contact_person": {
                    "type": "string",
                    "example": "Joan Laporta"
                },
                "contact_phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "123456789"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "FC Barcelona"
                }
            }
        },
        "team.MatchHistoryResponse": {
            "type": "object",
            "properties": {
                "draws": {
                    "type": "integer",
                    "example": 1
                },
                "losses": {
                    "type": "integer",
                    "example": 3
                },
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/team.MatchResponse"
                    }
                },
                "total_matches": {
                    "type": "integer",
                    "example": 10
                },
                "wins": {
                    "type": "integer",
                    "example": 6
                }
            }
        },
        "team.MatchResponse": {
            "type": "object",
            "properties": {
                "match_date": {
                    "type": "string",
                    "example": "2024-06-01T15:00:00Z"
                },
                "match_id": {
                    "type": "integer",
                    "example": 1
                },
                "opponent_score": {
                    "type": "integer",
                    "example": 1
                },
                "our_score": {
                    "type": "integer",
                    "example": 3
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "venue": {
                    "type": "string",
                    "example": "Camp Nou"
                }
            }
        },
        "team.TeamResponse": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "FC Barcelona Sports Club"
                },
                "contact_email": {
                    "type": "string",
                    "example": "contact@fcbarcelona.com"
                },
                "contact_person": {
                    "type": "string",
                    "example": "Joan Laporta"
                },
                "contact_phone": {
                    "type": "string",
                    "example": "123456789"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "FC Barcelona"
                }
            }
        },
        "team.TeamWithStatsResponse": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "FC Barcelona Sports Club"
                },
                "contact_email": {
                    "type": "string",
                    "example": "contact@fcbarcelona.com"
                },
                "contact_person": {
                    "type": "string",
                    "example": "Joan Laporta"
                },
                "contact_phone": {
                    "type": "string",
                    "example": "123456789"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "match_history": {
                    "$ref": "#/definitions/team.MatchHistoryResponse"
                },
                "name": {
                    "type": "string",
                    "example": "FC Barcelona"
                }
            }
        },
        "team.UpdateTeamRequest": {
            "type": "object",
            "required": [
                "company_name",
                "contact_person",
                "name"
            ],
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "FC Barcelona Sports Club Updated"
                },
                "contact_email": {
                    "type": "string",
                    "example": "updated@fcbarcelona.com"
                },
                "contact_person": {
                    "type": "string",
                    "example": "Updated Contact Person"
                },
                "contact_phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "987654321"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "FC Barcelona Updated"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}